big makefiles fix

+++ get_next_line/srcs/get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: sgertrud <msnazarow@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/29 12:21:57 by sgertrud          #+#    #+#             */
-/*   Updated: 2020/12/27 10:49:26 by sgertrud         ###   ########.fr       */
+/*   Updated: 2020/12/30 08:56:55 by sgertrud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,9 +94,9 @@ int			get_next_line(int fd, char **line)
 	if (!reminder || fd >= (long)reminder[0])
 	{
 		if (!(reminder = ft_realloc_2(reminder, reminder ?
-		(long)reminder[0] + 1 : FD + 1, fd + 10)))
+		(unsigned long)reminder[0] + 1 : FD + 1, (unsigned long)fd + 10)))
 			return (-1);
-		reminder[0] = (char*)((long)fd + 9);
+		reminder[0] = (char*)((unsigned long)fd + 9);
 		reminder[fd + 1] = 0;
 	}
 	out = g_line(fd, reminder, line);
